Add precision to BlockBox definition

+++ src/game/collision/CustomBlockBox.java
@@ -5,12 +5,12 @@ public class CustomBlockBox {
     private static final double[] aabb = new double[6];
 
     public static void setBlockBox(int x, int y, int z, double[] blockBox){
-        aabb[0] = x + blockBox[0];
-        aabb[1] = y + blockBox[1];
-        aabb[2] = z + blockBox[2];
-        aabb[3] = x + blockBox[3];
-        aabb[4] = y + blockBox[4];
-        aabb[5] = z + blockBox[5];
+        aabb[0] = (double)x + blockBox[0];
+        aabb[1] = (double)y + blockBox[1];
+        aabb[2] = (double)z + blockBox[2];
+        aabb[3] = (double)x + blockBox[3];
+        aabb[4] = (double)y + blockBox[4];
+        aabb[5] = (double)z + blockBox[5];
     }
 
     //getters
